MAINT: Rename cpdf to pdfly
Closes #12
