+++ .github/workflows/github-ci.yaml
@@ -32,7 +32,7 @@ jobs:
       run: |
         pip install -r requirements/ci.txt
 
-    - name: Install cpdf
+    - name: Install pdfly
       run: |
         pip install .
