+++ README.md
@@ -1,27 +1,27 @@
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
-[![PyPI version](https://badge.fury.io/py/cpdf.svg)](https://pypi.org/project/cpdf/)
-[![GitHub last commit](https://img.shields.io/github/last-commit/py-pdf/cpdf)](https://github.com/py-pdf/cpdf)
-[![Python Support](https://img.shields.io/pypi/pyversions/cpdf.svg)](https://pypi.org/project/cpdf/)
+[![PyPI version](https://badge.fury.io/py/pdfly.svg)](https://pypi.org/project/pdfly/)
+[![GitHub last commit](https://img.shields.io/github/last-commit/py-pdf/pdfly)](https://github.com/py-pdf/pdfly)
+[![Python Support](https://img.shields.io/pypi/pyversions/pdfly.svg)](https://pypi.org/project/pdfly/)
 
-# cpdf
+# pdfly
 
-cpdf is a pure-python cli application for manipulating PDF files.
+pdfly (say: PDF-li) is a pure-python cli application for manipulating PDF files.
 
 ## Installation
 
 ```bash
-pip install -U cpdf
+pip install -U pdfly
 ```
 
-As `cpdf` is an application, you might want to install it with [`pipx`](https://pypi.org/project/pipx/).
+As `pdfly` is an application, you might want to install it with [`pipx`](https://pypi.org/project/pipx/).
 
 ## Usage
 
 ```console
-$ cpdf --help
-Usage: cpdf [OPTIONS] COMMAND [ARGS]...
+$ pdfly --help
+Usage: pdfly [OPTIONS] COMMAND [ARGS]...
 
-  cpdf is a pure-python cli application for manipulating PDF files.
+  pdfly is a pure-python cli application for manipulating PDF files.
 
 Options:
   --version
@@ -40,8 +40,8 @@ Commands:
 You can see the help of every subcommand by typing:
 
 ```console
-$ cpdf 2-up --help
-Usage: cpdf 2-up [OPTIONS] PDF OUT
+$ pdfly 2-up --help
+Usage: pdfly 2-up [OPTIONS] PDF OUT
 
   Create a booklet-style PDF from a single input.
