--- cpdf/__main__.py
@@ -1,6 +1,0 @@
-"""Execute cpdf as a module."""
-
-from cpdf.cli import entry_point
-
-if __name__ == "__main__":
-    entry_point()
