+++ pdfly/__main__.py
@@ -1,0 +1,6 @@
+"""Execute pdfly as a module."""
+
+from pdfly.cli import entry_point
+
+if __name__ == "__main__":
+    entry_point()
