+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 # https://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
-name = cpdf
+name = pdfly
 
 author = Martin Thoma
 author_email = info@martin-thoma.de
@@ -12,7 +12,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 keywords = pdf, pure-python
 
-url = https://github.com/py-pdf/cpdf
+url = https://github.com/py-pdf/pdfly
 license = BSD-3-Clause
 license_file = LICENSE
 
@@ -42,7 +42,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    cpdf = cpdf.cli:entry_point
+    pdfly = pdfly.cli:entry_point
 
 [tool:pytest]
 addopts = --doctest-modules --cov=. --cov-report html:tests/reports/coverage-html --cov-report term-missing --ignore=docs/ --durations=3 --timeout=30
